Partial rewrite of API and Testing

+++ featurerequest/apiviews.py
@@ -1,7 +1,7 @@
 """Provides all RESTful API calls via Flask-RESTful"""
 from datetime import datetime
 from featurerequest import app, db
-from featurerequest.user_auth import roles
+from featurerequest.user_auth import roles, roles_advanced
 from featurerequest.models import \
     User, UserSchema,\
     Client, ClientSchema,\
@@ -19,8 +19,91 @@ api = Api(app) # pylint: disable=invalid-name
 
 # Used to prevent the RESTful methods from being flagged.
 # pylint: disable=no-self-use
+class BaseAPI(Resource):
+    """Provides a custom base for all resources to use, helping to minimize code use."""
 
-class UserAPI(Resource):
+    def __init__(self, *args, **kwargs):
+        super(Resource, self).__init__(*args, **kwargs)
+        self.resource_url = ''
+        self.role_list = {
+            'get': [],
+            'post': [],
+            'put': [],
+            'delete': []
+        }
+        self.resource_name = ''
+        self.single_item = ''
+        self.multiple_items = ''
+        self.table = None
+        self.schema = None
+        self.schema_many = None
+        self.put_check_message = ''
+
+
+
+    @roles_advanced('get')
+    def get(self, id=None):
+        """Returns a list of all resources ir the resource belonging to the id."""
+        if id is None:
+            query = self.table.query.all()
+            results = self.schema_many.dump(query)
+            return make_response(jsonify({self.multiple_items: results.data}), 200)
+        query = self.table.query.get_or_404(id)
+        result = self.schema.dump(query)
+        return make_response(jsonify({self.single_item: result.data}), 200)
+
+    @roles_advanced('post')
+    def post(self):
+        """Adds a resource and returns the added database entry."""
+        json_data = request.get_json()
+        if not json_data:
+            return make_response(jsonify({'message': 'No data provided'}), 400)
+        data, errors = self.schema.load(json_data)
+        if errors:
+            return make_response(jsonify(errors), 400)
+        data.user_id = current_user.id
+        db.session.add(data)
+        try:
+            db.session.commit()
+        except IntegrityError:
+            db.session.rollback()
+            return make_response(jsonify({'message': self.resource_name +
+                                          ' is already in the database!'}), 400)
+        result = self.schema.dump(self.table.query.get_or_404(data.id))
+        return make_response(jsonify({'message': 'Added the ' + self.resource_name + '.',
+                                     self.single_item: result.data}), 201)
+
+    @roles_advanced('put')
+    def put(self, id=None):
+        """Updates a resource and returns the updated entry."""
+        if not id:
+            return make_response(jsonify({
+                'message': 'You must supply an id in the URI.'}), 400)
+        try:
+            safe = self.put_check(id)
+            if not safe:
+                return make_response(jsonify({
+                'message': self.put_check_message}), 400)
+        except AttributeError: # Happens if self.put_check() isn't defined. EAFP!
+            pass
+        json_data = request.get_json()
+        if not json_data:
+            return make_response(jsonify({'message': 'No data provided'}), 400)
+        data = self.table.query.get_or_404(id)
+        try:
+            for key, value in json_data.items():
+                self.schema.validate({key:value})
+                setattr(data, key, value)
+                db.session.commit()
+        except ValidationError as err:
+            db.session.rollback()
+            return make_response(jsonify({'message': err.messages}), 400)
+        result = self.schema.dump(self.table.query.get_or_404(data.id))
+        return make_response(jsonify({'message': 'Updated the ' + self.resource_name + '.',
+                                      self.single_item: result.data}), 200)
+
+
+class UserAPI(BaseAPI):
     """Provides the 'user' resource.
 
     get/<user_id>:
@@ -35,73 +118,44 @@ class UserAPI(Resource):
     delete/<user_id>:
         Sets the specified user to 'Inactive', preventing them from using the service.
     """
-    @roles('Employee', 'Administrator')
-    def get(self, user_id=None):
-        """Return a list of users or the user belonging to the user_id."""
-        if user_id is None:
-            users = User.query.all()
-            results = UserSchema(many=True).dump(users)
-            return make_response(jsonify({'users': results.data}), 200)
-        user = User.query.get(user_id)
-        result = UserSchema().dump(user)
-        return make_response(jsonify({'user': result.data}), 200)
+    def __init__(self, *args, **kwargs):
+        super(BaseAPI, self).__init__(*args, **kwargs)
+        self.resource_url = ''
+        self.role_list = {
+            'get': ['Employee', 'Administrator'],
+            'post': ['Administrator'],
+            'put': ['Administrator'],
+            'delete': ['Administrator']
+        }
+        self.resource_name = 'User'
+        self.single_item = 'user'
+        self.multiple_items = 'users'
+        self.table = User
+        self.schema = UserSchema()
+        self.schema_many = UserSchema(many=True)
+        self.put_check_message = 'You cannot modify your own attributes.'
+
+    def put_check(self, id):
+        if current_user.id == id:
+            return False
+        return True
 
     @roles('Administrator')
-    def post(self):
-        """Adds a new user, returns the added database entry."""
-        schema = UserSchema()
-        json_data = request.get_json()
-        if not json_data:
-            return make_response(jsonify({'message': 'No data provided'}), 400)
-        data, errors = schema.load(json_data)
-        if errors:
-            return make_response(jsonify(errors), 422)
-        data.user_id = current_user.id
-        db.session.add(data)
-        try:
-            db.session.commit()
-        except IntegrityError:
-            return make_response(jsonify({'message': 'Username already in the database!'}), 400)
-        print(data.username)
-        result = schema.dump(User.query.get(data.id))
-        print(result)
-        return make_response(jsonify({'message': 'Added a user.', 'user': result.data}), 201)
-
-    @roles('Administrator')
-    def put(self, user_id):
-        """Updates the user."""
-        if current_user.id == user_id:
+    def delete(self, id=None):
+        if not id:
             return make_response(jsonify({
                 'message': 'You cannot modify your own attributes.'}), 400)
-        schema = UserSchema()
-        json_data = request.get_json()
-        if not json_data:
-            return make_response(jsonify({'message': 'No data provided'}), 400)
-        data = User.query.get_or_404(json_data['id'])
-        try:
-            for key, value in json_data.items():
-                schema.validate({key:value})
-                setattr(data, key, value)
-                db.session.commit()
-        except ValidationError as err:
-            return make_response(jsonify({'message': err.messages}), 400)
-        result = schema.dump(User.query.get(data.id))
-        print(result)
-        return make_response(jsonify({'message': 'Updated a user.', 'user': result.data}), 201)
-
-    @roles('Administrator')
-    def delete(self, user_id):
         """Sets the specified user to 'Inactive', preventing them from using the service."""
-        if current_user.id == user_id:
+        if current_user.id == id:
             return make_response(jsonify({
                 'message': 'You cannot modify your own attributes.'}), 400)
-        data = User.query.get_or_404(user_id)
+        data = User.query.get_or_404(id)
         data.role = 'Inactive'
         db.session.commit()
         return make_response(jsonify({'message': 'Set the user to inactive.',
                                       'user': {'username': data.username}}), 200)
 
-class ClientAPI(Resource):
+class ClientAPI(BaseAPI):
     """Provides the 'client' resource.
 
     get/<client_id>:
@@ -116,64 +170,37 @@ class ClientAPI(Resource):
     delete/<client_id>:
         Remove the client and all data associated with it.
     """
-    @roles('Employee', 'Administrator')
-    def get(self, client_id=None):
-        """Return a list of clients or the client belonging to the client_id."""
-        if client_id is None:
-            clients = Client.query.all()
-            results = ClientSchema(many=True).dump(clients)
-            return make_response(jsonify({'clients': results.data}), 200)
-        client = Client.query.get(client_id)
-        result = ClientSchema().dump(client)
-        return make_response(jsonify({'client': result.data}), 200)
+    def __init__(self, *args, **kwargs):
+        super(BaseAPI, self).__init__(*args, **kwargs)
+        self.resource_url = ''
+        self.role_list = {
+            'get': ['Employee', 'Administrator'],
+            'post': ['Administrator'],
+            'put': ['Administrator'],
+            'delete': ['Administrator']
+        }
+        self.resource_name = 'Client'
+        self.single_item = 'client'
+        self.multiple_items = 'clients'
+        self.table = Client
+        self.schema = ClientSchema()
+        self.schema_many = ClientSchema(many=True)
 
     @roles('Administrator')
-    def post(self):
-        """Adds a new client, returns the added database entry."""
-        schema = ClientSchema()
-        json_data = request.get_json()
-        if not json_data:
-            return make_response(jsonify({'message': 'No data provided'}), 400)
-        data, errors = schema.load(json_data)
-        if errors:
-            return make_response(jsonify(errors), 422)
-        data.user_id = current_user.id
-        db.session.add(data)
-        db.session.commit()
-        result = schema.dump(Client.query.get(data.id))
-        return make_response(jsonify({'message': 'Added a client.', 'client': result.data}), 201)
-
-    @roles('Administrator')
-    def put(self, client_id):
-        """Updates the client."""
-        schema = ClientSchema()
-        json_data = request.get_json()
-        if not json_data:
-            return make_response(jsonify({'message': 'No data provided'}), 400)
-        data = Client.query.get_or_404(client_id)
-        try:
-            for key, value in json_data.items():
-                schema.validate({key:value})
-                setattr(data, key, value)
-                db.session.commit()
-        except ValidationError as err:
-            return make_response(jsonify({'message': err.messages}), 400)
-        result = schema.dump(Client.query.get(data.id))
-        print(result)
-        return make_response(jsonify({'message': 'Updated a client.', 'client': result.data}), 201)
-
-    @roles('Administrator')
-    def delete(self, client_id):
+    def delete(self, id=None):
         """Remove the client and all data associated with it."""
-        data = Client.query.get_or_404(client_id)
+        if not id:
+            return make_response(jsonify({
+                'message': 'You must supply an id in the URI.'}), 400)
+        data = Client.query.get_or_404(id)
         name = data.name
         # Remove all dependencies
         feature_ids = [x[0] for x in db.session.query(Feature.id).\
-            filter_by(client_id=client_id).all()]
+            filter_by(client_id=id).all()]
         for feature_id in feature_ids:
             db.session.query(FeatureComment).filter_by(feature_id=feature_id).delete()
-        db.session.query(Feature).filter_by(client_id=client_id).delete()
-        db.session.query(Client).filter_by(id=client_id).delete()
+        db.session.query(Feature).filter_by(client_id=id).delete()
+        db.session.query(Client).filter_by(id=id).delete()
         db.session.commit()
         return make_response(jsonify({'message': 'Removed a client.',
                                       'client': {'name': name}}), 200)
@@ -239,9 +266,8 @@ class ProductAPI(Resource):
         except ValidationError as err:
             return make_response(jsonify({'message': err.messages}), 400)
         result = schema.dump(Product.query.get(data.id))
-        print(result)
         return make_response(jsonify({'message': 'Updated a product.',
-                                      'product': result.data}), 201)
+                                      'product': result.data}), 200)
 
     @roles('Administrator')
     def delete(self, product_id):
@@ -287,11 +313,12 @@ class FeatureAPI(Resource):
         json_data = request.get_json()
         if not json_data:
             return make_response(jsonify({'message': 'No data provided'}), 400)
-        if json_data['target_date'] != "":
-            json_data['target_date'] = str(datetime.strptime(json_data['target_date'],
-                                                             "%m/%d/%Y").strftime("%Y-%m-%d"))
-        else:
-            del json_data['target_date']
+        if 'target_date' in json_data:
+            if json_data['target_date'] != "":
+                json_data['target_date'] = str(datetime.strptime(json_data['target_date'],
+                                                                 "%m/%d/%Y").strftime("%Y-%m-%d"))
+            else:
+                del json_data['target_date']
         data, errors = schema.load(json_data)
         if errors:
             return make_response(jsonify(errors), 422)
@@ -331,7 +358,6 @@ class FeatureAPI(Resource):
         data = Feature.query.get_or_404(feature_id)
         priority = int(json_data['priority'])
         if priority != data.priority:
-            print("Entering")
             # Changing the priority... Lets move everything down if needed.
             results = Feature.query.filter(Feature.client_id == json_data['client_id']).\
                 filter(Feature.priority > data.priority).all()
@@ -347,7 +373,6 @@ class FeatureAPI(Resource):
                 db.session.commit()
             results = db.session.query(func.max(Feature.priority)).\
                 filter(Feature.client_id == int(json_data['client_id'])).all()
-            print(results)
             max_priority = results[0][0]
             if max_priority is None:
                 json_data['priority'] = 1
@@ -363,9 +388,8 @@ class FeatureAPI(Resource):
         except ValidationError as err:
             return make_response(jsonify({'message': err.messages}), 400)
         result = schema.dump(Feature.query.get(data.id))
-        print(result)
         return make_response(jsonify({'message': 'Updated a feature.',
-                                      'feature': result.data}), 201)
+                                      'feature': result.data}), 200)
 
     @roles('Employee', 'Administrator')
     def delete(self, feature_id):
@@ -425,8 +449,8 @@ class FeatureCommentAPI(Resource):
                                       'feature_note': result.data}), 201)
 
 
-api.add_resource(UserAPI, '/user', '/user/<int:user_id>', endpoint='user')
-api.add_resource(ClientAPI, '/client', '/client/<int:client_id>', endpoint='client')
+api.add_resource(UserAPI, '/user', '/user/<int:id>', endpoint='user')
+api.add_resource(ClientAPI, '/client', '/client/<int:id>', endpoint='client')
 api.add_resource(ProductAPI, '/product', '/product/<int:product_id>', endpoint='product')
 api.add_resource(FeatureAPI, '/feature', '/feature/<int:feature_id>', endpoint='feature')
 api.add_resource(FeatureCommentAPI, '/feature_comment', '/feature_comment/<int:feature_id>',
